+++ buildSrc/src/main/resources/META-INF/gradle-plugins/carrotsearch.randomized-testing.properties
@@ -1,0 +1,1 @@
+implementation-class=com.carrotsearch.gradle.junit4.RandomizedTestingPlugin
