--- buildSrc/src/main/resources/META-INF/gradle-plugins/carrotsearch.randomizedtesting.properties
@@ -1,1 +1,0 @@
-implementation-class=com.carrotsearch.gradle.randomizedtesting.RandomizedTestingPlugin
